+++ view/ace.go
@@ -34,7 +34,7 @@ func (s *AceEngine) SetIndent(indent string) *AceEngine {
 // Usage:
 // Ace("./views", ".ace") or
 // Ace(iris.Dir("./views"), ".ace") or
-// Ace(AssetFile(), ".ace") for embedded data.
+// Ace(embed.FS, ".ace") or Ace(AssetFile(), ".ace") for embedded data.
 func Ace(fs interface{}, extension string) *AceEngine {
 	s := &AceEngine{HTMLEngine: HTML(fs, extension), indent: ""}
 	s.name = "Ace"